move mock function file
