+++ examples/main/shared/utils/linkPreviewUtil.js
